refactor(ladderPath): LadderPath enum 패키지 변경
- /constants -> /model
