+++ src/main/java/ladder/model/LadderPath.java
@@ -1,4 +1,4 @@
-package ladder.constant;
+package ladder.model;
 
 public enum LadderPath {
     STAY, LEFT, RIGHT;
